Implement initial SLE support
SLE exports from SUSE Studio replace the repository URLS with placeholders which
we need to change to OBS repositories back.
Furthermore we need to add a repository to substitute some dependencies.
For now we replace all placeholders with the SP3 projects in OBS as
we don't support older than SLE12 SP3 builds in OBS for now.